Add files via upload
Hello friends. This is my first project in C and it's a typical airport project. It contains a lot of things like in real life, that is, it is a simulation. They are divided into two as the Company and the Passenger side, and they are divided into sections within themselves. For continuity, every time the code runs, it adds a new day file (22.02.2022) and adds the Ports in it.



Company Features
-The free ports on the desired date are checked and presented.

-When the company chooses the desired port, the aircraft shape is written on the port and the empty seats are indicated with an 'O'.

-The company can check how many planes it has at any time (the control function compares it with the name in all txt files).



Passenger Features

-Passenger can see the ports with aircraft on the desired date.

-When the seat is selected, the seat (7C) chosen by the passenger turns to 'X' in the txt file.

- Passenger information can check-in.



Generally

-Passenger information is edited and written into the file of that date.

-A txt file visible to the management shows all passengers at the airport.